Add files via upload
The project is a portable telematics device that collects real-time vehicle data. Using SPI, UART, and Bluetooth, it captures and transmits NO2 emissions, speed, and x, y, z axis and temperature data. This information is securely transmitted to the cloud. The system runs on FreeRTOS for efficient data acquisition. This will allows to notice the driving behavior and 
environmental impact.